[optin/reporting] task: moved shared run* file to test/test
