+++ optin/reporting/test/models/completed.spec.js
@@ -1,5 +1,5 @@
-const runDBNotFoundTests = require('./runDBNotFoundTests');
-const runTotalRecordTests = require('./runTotalRecordTests');
+const runDBNotFoundTests = require('../test/runDBNotFoundTests');
+const runTotalRecordTests = require('../test/runTotalRecordTests');
 const Config = require('../../config');
 const completedDB = require('../../models/completed');
 const { refreshDBs } = require('../../models/database');
